% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{coef.glmnet_h2o4gpu}
\alias{coef.glmnet_h2o4gpu}
\title{glmnet_h2o4gpu coefficients.}
\usage{
\method{coef}{glmnet_h2o4gpu}(object, s = NULL, ...)
}
\arguments{
\item{object}{Value returned by h2o4gpu::glmnet(..).}

\item{s}{Values of lambda for which coefficient should be returned.}

\item{...}{Not used. Other arguments to predict.}
}
\description{
Returns coeficients of glmnet_h2o4gpu fit.
}
