% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{coef.h2o4gpunet}
\alias{coef.h2o4gpunet}
\title{h2o4gpunet coefficients.}
\usage{
\method{coef}{h2o4gpunet}(object, s = NULL, ...)
}
\arguments{
\item{object}{Value returned by h2o4gpunet(..).}

\item{s}{Values of lambda for which coefficient should be returned.}

\item{...}{Not used. Other arguments to predict.}
}
\description{
Returns coeficients of h2o4gpunet fit.
}
