% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{cv.h2o4gpunet}
\alias{cv.h2o4gpunet}
\title{Cross-validation for glmnet}
\usage{
cv.h2o4gpunet(x, y, weights, lambda = NULL, nfolds = 10, foldid = NULL,
  noweight = FALSE, cutoff = TRUE, params = list(), ...)
}
\arguments{
\item{x}{Data matrix.}

\item{y}{Response vector}

\item{weights}{Weights of each training example, defaults to 1.}

\item{lambda}{List of lambda values for which CV should be performed.}

\item{nfolds}{Number of folds.}

\item{foldid}{Optional vector of values between 1 and nfolds, assigning}

\item{cutoff}{Discard values of lambda for which beta remains unchanged (estimated on the full dataset)
datapoint to cv-fold.}

\item{params}{Other parameters to h2o4gpunet}

\item{...}{Other parameters to h2o4gpunet.}
}
\description{
Does k-fold cross-validation for glmnet.
}
