% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{kIndBox01}
\alias{kIndBox01}
\title{Indicator Function of [0, 1] Box}
\usage{
kIndBox01(m = 1)
}
\arguments{
\item{m}{Number of times that the function should be repeated.}
}
\description{
\deqn{h(x) = I(0 <= x <= 1)}
}
