% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{kIndEq0}
\alias{kIndEq0}
\title{Indicator Function of Origin}
\usage{
kIndEq0(m = 1)
}
\arguments{
\item{m}{Number of times that the function should be repeated.}
}
\description{
\deqn{h(x) = I(x = 0)}
}
