% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2o4gpu.R
\name{kIndGe0}
\alias{kIndGe0}
\title{Indicator Function of Non-Negative Orthant}
\usage{
kIndGe0(m = 1)
}
\arguments{
\item{m}{Number of times that the function should be repeated.}
}
\description{
\deqn{h(x) = I(x >= 0)}
}

