% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{plot.cv.glmnet_h2o4gpu}
\alias{plot.cv.glmnet_h2o4gpu}
\title{Plot H2O4GPU GLMNET CV}
\usage{
\method{plot}{cv.glmnet_h2o4gpu}(x, sign.lambda, ...)
}
\arguments{
\item{x}{cv.glmnet_h2o4gpu object}

\item{sign.lambda}{Either plot against log(lambda) (default) or its negative if sign.lambda=-1.}

\item{...}{Other graphical parameters to plot.}
}
\description{
Plots results of glmnet CV
}
