% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{plot.cv.h2o4gpunet}
\alias{plot.cv.h2o4gpunet}
\title{Plot h2o4gpunet CV}
\usage{
\method{plot}{cv.h2o4gpunet}(x, sign.lambda, ...)
}
\arguments{
\item{x}{h2o4gpunet.cv object}

\item{sign.lambda}{Either plot against log(lambda) (default) or its negative if sign.lambda=-1.}

\item{...}{Other graphical parameters to plot.}
}
\description{
Plots results of h2o4gpunet CV
}
