% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{plot.glmnet_h2o4gpu}
\alias{plot.glmnet_h2o4gpu}
\title{glmnet_h2o4gpu coefficient plot.}
\usage{
\method{plot}{glmnet_h2o4gpu}(x, xvar = c("norm", "lambda", "dev"), ...)
}
\arguments{
\item{x}{Value returned by h2o4gpu::glmnet(..).}

\item{xvar}{x-variable to plot coefficient against.}

\item{...}{Extra parameters for matplot.}
}
\description{
Plots coefficients
}
