% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/h2o4gpu.R
\name{plot.h2o4gpunet}
\alias{plot.h2o4gpunet}
\title{h2o4gpunet coefficient plot.}
\usage{
\method{plot}{h2o4gpunet}(x, xvar = c("norm", "lambda", "dev"), ...)
}
\arguments{
\item{x}{h2o4gpunet object.}

\item{xvar}{x-variable to plot coefficient against.}

\item{...}{Extra parameters for matplot.}
}
\description{
Plots coefficients
}

