% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{predict.h2o4gpunet}
\alias{predict.h2o4gpunet}
\title{H2O4GPUnet prediction}
\usage{
\method{predict}{h2o4gpunet}(object, newx, s = NULL, type = c("link",
  "response", "class"), ...)
}
\arguments{
\item{object}{Value returned by h2o4gpunet(..).}

\item{newx}{New value of x for which prediction should be done.}

\item{s}{Values of lambda for which coefficient should be returned.}

\item{type}{Predict response or class.}

\item{...}{Not used. Other arguments to predict.}
}
\description{
Predict response for new value input
}
