% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{print.glmnet_h2o4gpu}
\alias{print.glmnet_h2o4gpu}
\title{Prints glmnet_h2o4gpu summary}
\usage{
\method{print}{glmnet_h2o4gpu}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{Value returned by h2o4gpu::glmnet(..).}

\item{digits}{Number of digits to display.}

\item{...}{Not used.}
}
\description{
Prints glmnet_h2o4gpu summary
}
