% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o4gpuglm.R
\name{print.h2o4gpunet}
\alias{print.h2o4gpunet}
\title{Prints h2o4gpunet summary}
\usage{
\method{print}{h2o4gpunet}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{h2o4gpunet object}

\item{digits}{Number of digits to display}

\item{...}{Not used.}
}
\description{
Prings h2o4gpunet summary
}
